refactor(data): extraer las migraciones a db/migrations.cjs — fase b (05/T03)
Las 42 migraciones y sus 3 helpers (parseJsonValue, normalizeServerId,
extractLegacyMcpServers) se mueven verbatim a electron/core/db/migrations.cjs;
applyMigrations queda byte-idéntico (verificado por diff). database.cjs
baja de 3.971 a 1.097 líneas y conserva runMigrations (backup + restore
on failure) en la fachada, llamando al applyMigrations importado. Sin
cambios de comportamiento: instalación nueva = schema final + runner;
existente = runner.
